Refactored Phase Refactored methods to move and rotate convey belts to ConveyorBelt class.
